+++ json2ciw/__init__.py
@@ -1,17 +1,21 @@
 __author__ = "Tom Monks"
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 
 from .datasets import load_call_centre_model, load_model_file
 from .engine import CiwConverter, multiple_replications
-from .results import summarise_results
+from .results import summarise_results, tidy_to_wide_format, create_user_filtered_hist
+from .ui import render_simulation_app
 
 # define what is exported
 __all__ = [
     "load_model_file"
     "load_call_centre_model",
     "CiwConverter",
-    "multiple_replications"
-    "summarise_results"
+    "multiple_replications",
+    "summarise_results",
+    "render_simulation_app",
+    "tidy_to_wide_format",
+    "create_user_filtered_hist",
     "__version__",
     "__author__"
 ]